Refactor auth and user routes
Move ID data validation to middleware module

Configure express-session
