+++ routes/authRoutes.js
@@ -1,0 +1,39 @@
+const router = require('express').Router()
+
+// Import data models
+const db = require('../data/models')
+
+// Import middleware
+const { requiredData } = require('../middleware')
+
+// Authorization Resource Route
+router.post('/register', requiredData, async (req, res) => {
+  let user = req.body
+  const hash = bcrpyt.hashSync(user.password, 14)
+  user.password = hash
+
+  try {
+    let data = await db.insert(user, 'Users')
+    res.status(201).send(data)
+  }
+  catch (err) {
+    res.status(500).send(err.message)
+  }
+})
+
+router.post('/login', requiredData, async (req, res) => {
+  let { username, password } = req.body
+  try {
+    let user = await db.findByUser(username, 'Users')
+    if (user && bcrypt.compareSync(password, user.password)) {
+      res.json({ message: `Greetings ${username}!` })
+    } else {
+      res.status(401).json({ message: `You shall not pass!` })
+    }
+  }
+  catch (err) {
+    res.status(500).send(err.message)
+  }
+})
+
+module.exports = router
